Module rustyread::cli::simulate

All stuff relate to simulate subcommand

Structs

Command

Struct use to parse simulate subcommand argument

Duo

Store a pair of value, can be parse from str if it’s match with \d+,\d+

Quantity

Store quantity as coverage of number of base

Trio

Store a trio of value, can be parse from str if it’s match with \d+,\d+,\d++

Functions

found_model

Found path to model file