Struct rustyread::cli::simulate::Command

pub struct Command {

    pub reference_path: String,
    pub output_path: Option<String>,
    pub quantity: Quantity,
    pub length: Duo,
    pub identity: Trio,
    pub error_model: String,
    pub qscore_model: String,
    pub seed: Option<u64>,
    pub start_adapter: Duo,
    pub end_adapter: Duo,
    pub start_adapter_seq: String,
    pub end_adapter_seq: String,
    pub junk: f64,
    pub random: f64,
    pub chimera: f64,
    pub glitches: Trio,
    pub small_plasmid_bias: bool,
    pub nb_base_store: Option<Quantity>,

}

Struct use to parse simulate subcommand argument

Fields
reference_path: String

Path to reference fasta (can be gzipped, bzip2ped, xzped)

output_path: Option<String>

Path where read is write

quantity: Quantity

Either an absolute value (e.g. 250M) or a relative depth (e.g. 25x)

length: Duo

Fragment length distribution (mean and stdev)

identity: Trio

Sequencing identity distribution (mean, max and stdev)

error_model: String

Path to an error model file

qscore_model: String

Path to an quality score model file

seed: Option<u64>

Random number generator seed for deterministic output (default: different output each time)

start_adapter: Duo

Adapter parameters for read starts (rate and amount)

end_adapter: Duo

Adapter parameters for read ends (rate and amount)

start_adapter_seq: String

Adapter parameters for read starts

end_adapter_seq: String

Adapter parameters for read ends

junk: f64

This percentage of reads wil be low complexity junk

random: f64

This percentage of reads wil be random sequence

chimera: f64

Percentage at which separate fragments join together

glitches: Trio

Read glitch parameters (rate, size and skip)

small_plasmid_bias: bool

If set, then small circular plasmids are lost when the fragment length is too high (default: small plasmids are included regardless of fragment length)

nb_base_store: Option<Quantity>

Number of base, rustyread can store in ram before write in output in absolute value (e.g. 250M) or a relative depth (e.g. 25x)